python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas
scikit-learn
google-generativeai==0.3.2